add external RAM to memory map of Magnavox Odyssey 2
The Magnavox Odyssey 2 has internal and external RAMs, which are
accessed using unique instructions, instead of being mapped to a single
address space.

The o2em core used to only expose the internal RAM, but many (maybe
most?) games make use of the external RAM. Having no access to it is
problematic for achievement development, so the core has been updated
to expose the external RAM as `RC_MEMORY_TYPE_SYSTEM_RAM` right after
the internal RAM, creating a virtual memory map that only exists for
this purpose.

This commit updates the memory map in rcheevos accordingly.