Initialized byteswapped
While practically impossible, thanks to the return on line 795, in theory there is a code path that results in the byteswapped variable on line 810 of hash.c to be uninitialized when it is checked. The Dolphin cmake ninja compile path in particular throws a pedantic warning over it. This commit initializes it upon creation to a dummy value that will never get used but should be sufficient to satisfy any compilers.