Add rc_runtime_alloc
This function will allocate a rc_runtime_t instance, initialize it, and return it. If the allocation fails, it will return NULL.
rc_runtime_t has also been modified to add a boolean value to tell if the instance owns itself (i.e. was allocated in rc_runtime_alloc), and therefore should deallocate the instance in rc_runtime_destroy.
Accordingly, this boolean is set to true in rc_runtime_alloc. Implicitly, it is set to false in rc_runtime_init, so users should not use rc_runtime_init when using the rc_runtime_alloc API, else memory will leak.
This ensures that all current users will need to do no changes to their code, unless they explicitly want to use the new rc_runtime_alloc API.

The reasoning for adding this API is due to BizHawk's usage of this library. BizHawk is written in C#, which means heap memory is managed and the garbage collector may move memory around. This presents issues with rc_runtime as it will store pointers to itself. Such pointers may become invalid due to the struct, being placed on the managed heap, being moved around. This presents a clean and performant solution to this issue while also likely helping for future similar cases.