Add rc_runtime_set_userdata.
This creates a function which sets userdata for the runtime instance.
This userdata is used for all callbacks: the peek callback, the event handler callback, and the validate handler callback.
This also therefore changes how userdata is passed to the peek callback: it will not be passed alongside the peek callback, rather it is just derived from the rc_runtime_t instance.
(Although more internal calls which don't take in a rc_runtime_t will just pass the userdata alongside the peek callback, just as before).
This could alternatively be done by passing userdata alongside the callback (like peek callbacks were handled before), but userdata being different for all callbacks seems unlikely, and it is an optional component anyways, so it probably makes more sense to store it in rc_runtime_t and use it for all callbacks.