Add allprogress/hashlibrary endpoint wrappers
These functions when paired together allow an emulator to populate
its game list with the achievement count for each game, and the
number of achievements unlocked by the user.

NOTE: Consumers of these APIs should cache the responses to these
calls on disk. You should *not* issue an allprogress call every time
the application starts or the list refreshes.