Add 3DS memory map
The "real" addresses set here are based on 3DS virtual memory addresses, with this memory map intended to map 1:1 with 3DS userland virtual memory

Unsure if this would end up being desirable in practice. This has the benefit of making points map 1:1 to RetroAchievements memory space (which is a huge win given how pointer heavy 3DS games will be and the disjointed mess that "real" memory lies in the memory map). This will, however, balloon the memory space to search in for developers. BizHawk (the current 3DS RetroAchievements capable emulator), as an implementation detail, also slightly suffers for the achievement runtime as it will allocate 1GiB here for a memory block LUT; this is an implementation detail, however, just due to being a multi-system emulator, a standalone emulator wouldn't have such, and even so 1GiB of memory is relatively free for any machine that can handle 3DS emulation anyways.

A different memory map could potentially be made which only covers the "real" memory spaces, but doing so will still have a large amount of memory mapped (definitely more than 256MiB, the size of the New 3DS's FCRAM), and will be much harder for the current achievements implementation to handle pointers (given the reported problems with even handling MEM1/MEM2 on the Wii).